# packages/contracts/openapi/academic.json
{
  "components": {
    "schemas": {
...
      "ApiResponseMapStringObject": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "additionalProperties": {},
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/stats/attendance/dashboard": {
      "get": {
        "operationId": "dashboard",
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-stat-controller"
        ]
      }
    },
...
  }
}